feat: referendum summary
