--- src/renderer/features/governance/components/ReferendumDetails/index.ts
@@ -1,1 +1,0 @@
-export { ReferendumDetailsDialog } from './ReferendumDetailsDialog';
