+++ src/renderer/features/governance/components/VotingHistory/VotingHistoryLoadingScreen.tsx
@@ -1,0 +1,9 @@
+import { Loader } from '@shared/ui';
+
+export const VotingHistoryLoadingScreen = () => {
+  return (
+    <div className="flex items-center justify-center h-40 w-full">
+      <Loader color="primary" />
+    </div>
+  );
+};
