+++ src/renderer/shared/api/translation/locales/en.json
@@ -362,7 +362,12 @@
     "searchPlaceholder": "Search",
     "title": "Governance",
     "voteHistory": {
-      "title": "Vote history"
+      "listColumnAccount": "Account",
+      "listColumnVotingPower": "Voting power",
+      "listEmptyState": "No one voted here",
+      "title": "Vote history",
+      "totalVotesCount": "{value} votes {symbol}",
+      "totalVotesCountConviction": "{value} × {conviction}"
     },
     "voted": "Voted",
     "voting": {
